+++ src/themes/light.less
@@ -1,6 +1,5 @@
 @main: @white;
 @text: @white;
-@main-light: #181818;
 
 @body-background: @main;
 @menu-bg: @main;
